Rename Akka and Scala forkjoin modules
