+++ .github/scripts/check-javaagent-suppression-keys.sh
@@ -48,6 +48,12 @@ for file in $(find instrumentation -name "*Module.java"); do
 
   matches=$(perl -0 -ne "print if /$expected/" "$file" | wc -l)
   if [ "$matches" == 0 ]; then
+    if grep -q "expandDeprecatedNames" "$file" \
+      && grep -q "\"$simple_module_name|deprecated:" "$file" \
+      && { [ "$module_name" == "$simple_module_name" ] || grep -q "\"$module_name|deprecated:" "$file"; }
+    then
+      continue
+    fi
     echo "Expected to find $expected in $file"
     exit 1
   fi
