+++ instrumentation/scala-forkjoin-2.8/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/scala/forkjoin/v2_8/ScalaConcurrentInstrumentationModule.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.scala.forkjoin.v2_8;
+
+import static io.opentelemetry.javaagent.extension.instrumentation.internal.DeprecatedInstrumentationNames.expandDeprecatedNames;
+import static java.util.Arrays.asList;
+
+import com.google.auto.service.AutoService;
+import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
+import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
+import java.util.List;
+
+@AutoService(InstrumentationModule.class)
+public class ScalaConcurrentInstrumentationModule extends InstrumentationModule {
+  public ScalaConcurrentInstrumentationModule() {
+    super(
+        "scala-forkjoin",
+        expandDeprecatedNames(
+            "scala-forkjoin|deprecated:scala-fork-join",
+            "scala-forkjoin-2.8|deprecated:scala-fork-join-2.8"));
+  }
+
+  @Override
+  public List<TypeInstrumentation> typeInstrumentations() {
+    return asList(new ScalaForkJoinPoolInstrumentation(), new ScalaForkJoinTaskInstrumentation());
+  }
+}
