+++ settings.gradle.kts
@@ -183,7 +183,7 @@ include(":smoke-tests-otel-starter:spring-boot-reactive-common")
 
 include(":instrumentation:activej-http-6.0:javaagent")
 include(":instrumentation:akka:akka-actor-2.3:javaagent")
-include(":instrumentation:akka:akka-actor-fork-join-2.5:javaagent")
+include(":instrumentation:akka:akka-actor-forkjoin-2.5:javaagent")
 include(":instrumentation:akka:akka-http-10.0:javaagent")
 include(":instrumentation:alibaba-druid-1.0:javaagent")
 include(":instrumentation:alibaba-druid-1.0:library")
@@ -615,7 +615,7 @@ include(":instrumentation:rxjava:rxjava-3.1.1:javaagent")
 include(":instrumentation:rxjava:rxjava-3.1.1:library")
 include(":instrumentation:rxjava:rxjava-common-3.0:library")
 include(":instrumentation:rxjava:rxjava-common-3.0:testing")
-include(":instrumentation:scala-fork-join-2.8:javaagent")
+include(":instrumentation:scala-forkjoin-2.8:javaagent")
 include(":instrumentation:servlet:servlet-2.2:javaagent")
 include(":instrumentation:servlet:servlet-3.0:javaagent")
 include(":instrumentation:servlet:servlet-3.0:javaagent-testing")
